send email from the alias instead of the main account
The email provided in MAIL FROM is the main account's email address,
even if you provide a specific "from" json key in the API call, while
the FROM header is the email supplied via the api.  This commit causes
the MAIL FROM to match the supplied email and the FROM header.