Clean up after fixtures that modify the database
A lot of our tests depend on fixtures to insert some data into the database;
however, they don't then remove the data they added, affecting the
behavior of future tests. This diff attempts to fix that for a large number
of our test fixtures. I think it reduces test running time a tad since there
is less data that accumulates for the tests to have to deal with. Might just
be noise though.